hallelujah, I think I've got it mostly correct!
lots of changes.  Most importantly, when expanding the changes inside
an observance, instead of looping through rules, and then years, we
now loop through years and then rules.  This allows us to look at the
rules _in order_, and therefore always have the correct offset from
standard for the _previous_ rule, which allows us to calculate the UTC
datetime for the changeover properly.

Also fixed calculating local end times for a span.