Updated to 2009a.
Fixed a bug in the Olson parser that generated incorrect short names
when an zone used a ruleset that started _after_ the observance
started. The Olson database arbitrarily picks the first non-DST zone
in the rule (regardless of start date) so we do too.

Changed debugging in Olson parser to be controllable via a --debug
flag in parse_olson,