Handle the case where /etc/share/zoneinfo when a symbolic link
This becomes a problem when /etc/localtime is a copy, rather than a symbolic
link, and DateTime::TimeZone::Local::Unix must use File::Find to traverse all
the zone files to figure out which one was copied.

File::Find::find() does not follow symbolic links by default, though things
like -d do. So we resolve the link if necessary before calling find().

The only instance known to me that triggers this is testing under macOS 13
High Sierra. Actually running under that OS is fine, since /etc/localtime is a
symbolic link. But when testing, a copy is used.